% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decido-package.R
\docType{package}
\name{decido-package}
\alias{decido}
\alias{decido-package}
\title{decido: Bindings for 'Mapbox' Ear Cutting Triangulation Library}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
