% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dopey.cpp-package.R
\docType{package}
\name{dopey.cpp-package}
\alias{dopey.cpp}
\alias{dopey.cpp-package}
\title{dopey.cpp: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
