% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earcut.cpp-package.R
\docType{package}
\name{earcut.cpp-package}
\alias{earcut.cpp}
\alias{earcut.cpp-package}
\title{earcut.cpp: Minimal Rcpp Binding of Mapbox Earcut.hpp}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
