% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{earcut}
\alias{earcut}
\title{earcut}
\usage{
earcut(x, y, holes = -1L, numholes = 0L)
}
\arguments{
\item{x}{x-coordinate}

\item{y}{y-coordinate}

\item{holes}{index of starting position of each hole in x,y}

\item{numholes}{the number of holes}
}
\description{
earcut description
}
\details{
earcut details
}
\examples{
x <- c(0, 0, 0.75, 1, 0.5, 0.8, 0.69)
y <- c(0, 1, 1, 0.8, 0.7, 0.6, 0)
earcut(x, y)
## hole starts at 9
x <- c(0, 0, 0.75, 1, 0.5, 0.8, 0.69, 0, 0.2,
      0.5, 0.5, 0.3, 0.2, 0.2)
y <- c(0, 1, 1, 0.8, 0.7, 0.6, 0, 0, 0.2,
      0.2, 0.4, 0.6, 0.4, 0.2)
ind <- earcut(x, y, holes = 9, numholes = 1)
plot(x, y, asp = 1)
xy <- cbind(x, y)
apply(matrix(ind, 3), 2, function(i) polygon(xy[i + 1, ]))
## add another hole
x <- c(0, 0, 0.75, 1, 0.5, 0.8, 0.69, 0, 0.2,
      0.5, 0.5, 0.3, 0.2, 0.2,
      0.15, 0.23, 0.2)
y <- c(0, 1, 1, 0.8, 0.7, 0.6, 0, 0, 0.2,
      0.2, 0.4, 0.6, 0.4, 0.2,
      0.65, 0.65, 0.81)
ind <- earcut(x, y, holes = c(9, 14), numholes = 2)
plot(x, y, asp = 1)
xy <- cbind(x, y)
apply(matrix(ind, 3), 2, function(i) polygon(xy[i + 1, ]))
}
