% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earcut.R
\name{plot_ears}
\alias{plot_ears}
\alias{plot_ears}
\alias{plot_holes}
\title{Plot ears or polygons}
\usage{
plot_ears(xy, idx, add = FALSE, ...)

plot_holes(xy, holes = 0, add = FALSE, ...)
}
\arguments{
\item{xy}{matrix of x, y coordinates}

\item{idx}{index of triangles}

\item{...}{arguments to polypath}

\item{holes}{index of starting position of holes (see `earcut`)}
}
\description{
Plot the triangles for earcut.
}
