% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decido-package.R
\docType{package}
\name{decido-package}
\alias{decido}
\alias{decido-package}
\alias{decido-package}
\title{decido: Bindings for 'Mapbox' Ear Cutting Triangulation Library}
\description{
Provides constrained triangulation of polygons. Ear cutting (or
ear clipping) applies constrained triangulation by successively 'cutting'
triangles from a polygon defined by path/s. Holes are supported by introducing
a bridge segment between polygon paths. This package wraps the 'header-only'
library 'earcut.hpp' <https://github.com/mapbox/earcut.hpp.git>.

Provides constrained triangulation of polygons. Ear cutting (or
ear clipping) applies constrained triangulation by successively 'cutting'
triangles from a polygon defined by path/s. Holes are supported.
This package wraps the 'header-only' library 'earcut.hpp'
\url{https://github.com/mapbox/earcut.hpp.git}.
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

Other contributors:
\itemize{
  \item Andrew Smith (provided C++ guidance) [contributor]
  \item Mapbox (author of header library earcut.hpp) [copyright holder]
}

}
\keyword{internal}
