% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{earcut}
\alias{earcut}
\title{earcut}
\usage{
earcut(x, y)
}
\arguments{
\item{x}{-ordinate}

\item{y}{-ordinate}
}
\description{
earcut description
}
\details{
earcut details
}
\examples{
x <- c(0, 0, 0.75, 1, 0.5, 0.8, 0.69)
y <- c(0, 1, 1, 0.8, 0.7, 0.6, 0)
earcut(x, y)
}
