% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earcut.R
\name{earcut}
\alias{earcut}
\title{earcut}
\usage{
earcut(x, y, holes = 0, ...)
}
\arguments{
\item{x}{x-coordinate}

\item{y}{y-coordinate}

\item{holes}{index of starting position of each hole in x,y, leave set to `0` if no holes}
}
\description{
earcut description
}
\details{
earcut details
}
\examples{
x <- c(0, 0, 0.75, 1, 0.5, 0.8, 0.69)
y <- c(0, 1, 1, 0.8, 0.7, 0.6, 0)
(ind <- earcut(x, y))
plot_ears(cbind(x, y), ind)
x <- c(0, 0, 0.75, 1, 0.5, 0.8, 0.69,
     0.2, 0.5, 0.5, 0.3, 0.2)
y <- c(0, 1, 1, 0.8, 0.7, 0.6, 0,
     0.2, 0.2, 0.4, 0.6, 0.4)
ind <- earcut(x, y, holes = 8)
plot_ears(cbind(x, y), ind)
## add another hole
x <- c(0, 0, 0.75, 1, 0.5, 0.8, 0.69,
     0.2, 0.5, 0.5, 0.3, 0.2,
      0.15, 0.23, 0.2)
y <- c(0, 1, 1, 0.8, 0.7, 0.6, 0,
     0.2, 0.2, 0.4, 0.6, 0.4,
      0.65, 0.65, 0.81)
ind <- earcut(x, y, holes = c(8, 13))
plot_ears(cbind(x, y), ind, col = "grey")
## more than one hole
## but the two inside holes are open to each other
## (so we can use the same data for one hole or two)
x <- c(0, 0, 1, 1,
       0.4, 0.2, 0.2, 0.4,
       0.6, 0.8, 0.8, 0.6
)
y <- c(0, 1, 1, 0,
       0.2, 0.2, 0.4, 0.4,
       0.6, 0.6, 0.4, 0.4
)
plot(x, y, type = "n")
ind <- decido::earcut(x, y, holes = c(5, 9))
plot_ears(cbind(x, y), ind, col = "grey")
plot_holes(cbind(x, y), holes = c(5, 9), col = "grey")
ind <- decido::earcut(x, y, holes = 5)
plot_ears(cbind(x, y), ind, col = "grey")
plot_holes(cbind(x, y), holes = 5, col = "grey")
}
